api-ai/requirements.txt
alembic==1.12.0 ; python_version >= "3.8" and python_version < "4.0"
alfred-cli==1.2.0 ; python_version >= "3.8" and python_version < "4.0"
aniso8601==9.0.1 ; python_version >= "3.8" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
decorator==5.1.1 ; python_version >= "3.8" and python_version < "4.0"
fake-store==0.2 ; python_version >= "3.8" and python_version < "4.0"
filelock==3.12.4 ; python_version >= "3.8" and python_version < "4.0"
flask-injector==0.14.0 ; python_version >= "3.8" and python_version < "4.0"
flask-restx==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
flask-sqlalchemy==3.0.3 ; python_version >= "3.8" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.8" and python_version < "4.0"
fsspec==2023.9.2 ; python_version >= "3.8" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.8" and python_version < "4.0"
honcho==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
huggingface-hub==0.17.2 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.8" and python_version < "3.10"
importlib-resources==6.1.0 ; python_version >= "3.8" and python_version < "3.9"
injector==0.21.0 ; python_version >= "3.8" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
jsonschema-specifications==2023.7.1 ; python_version >= "3.8" and python_version < "4.0"
jsonschema==4.19.1 ; python_version >= "3.8" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.8" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
networkx==3.1 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.8" and python_version < "4.0"
pillow==10.0.1 ; python_version >= "3.8" and python_version < "4.0"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
plumbum==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
psycopg2-binary==2.9.7 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==0.20.0 ; python_version >= "3.8" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.8" and python_version < "4.0"
pywin32==306 ; platform_system == "Windows" and platform_python_implementation != "PyPy" and python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.8" and python_version < "4.0"
referencing==0.30.2 ; python_version >= "3.8" and python_version < "4.0"
regex==2023.8.8 ; python_version >= "3.8" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.8" and python_version < "4.0"
retrying==1.3.4 ; python_version >= "3.8" and python_version < "4.0"
rpds-py==0.10.3 ; python_version >= "3.8" and python_version < "4.0"
safetensors==0.3.3 ; python_version >= "3.8" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
sqlalchemy-cockroachdb==1.4.4 ; python_version >= "3.8" and python_version < "4.0"
sqlalchemy==1.4.49 ; python_version >= "3.8" and python_version < "4.0"
sympy==1.12 ; python_version >= "3.8" and python_version < "4.0"
tokenizers==0.13.3 ; python_version >= "3.8" and python_version < "4.0"
torch==2.0.1 ; python_version >= "3.8" and python_version < "4.0"
torchvision==0.15.2 ; python_version >= "3.8" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.8" and python_version < "4.0"
transformers==4.33.2 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.8" and python_version < "4.0"
urllib3==2.0.5 ; python_version >= "3.8" and python_version < "4.0"
vcrpy==4.3.0 ; python_version >= "3.8" and python_version < "4.0"
werkzeug==2.3.7 ; python_version >= "3.8" and python_version < "4.0"
wrapt==1.15.0 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.8" and python_version < "3.10"
